% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{data}
\name{arrabida}
\alias{arrabida}
\title{Sample data of spiders in Arrabida (Portugal)}
\format{A data frame with 320 samples (rows) and 338 species (variables)}
\usage{
data(arrabida)
}
\description{
A dataset containing the abundance of 338 spider species in each of 320 samples. Details are described in:
Cardoso, P., Gaspar, C., Pereira, L.C., Silva, I., Henriques, S.S., Silva, R.R. & Sousa, P. (2008) Assessing spider species richness and composition in Mediterranean cork oak forests. Acta Oecologica, 33: 114-127.
}
\keyword{datasets}

