% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BAT.R
\docType{data}
\name{functree}
\alias{functree}
\title{Functional tree for 338 species of spiders}
\format{
An hclust object with 338 species.
}
\usage{
data(functree)
}
\description{
A dataset representing the functional tree for 338 species of spiders captured in Portugal.
For each species were recorded: average size, type of web, type of hunting, stenophagy, vertical stratification in vegetation and circadial activity. Details are described in:
Cardoso, P., Pekar, S., Jocque, R. & Coddington, J.A. (2011) Global patterns of guild composition and functional diversity of spiders. PLoS One, 6: e21710.
}
\keyword{datasets}
