% Generated by roxygen2 (4.0.0): do not edit by hand
\docType{data}
\name{geres}
\alias{geres}
\title{Sample data of spiders in Geres (Portugal)}
\format{A data frame with 320 rows and 338 variables}
\usage{
data(geres)
}
\description{
A dataset containing the abundance of 338 spider species in each of 320 samples. Details are described in:
Cardoso, P., Scharff, N., Gaspar, C., Henriques, S.S., Carvalho, R., Castro, P.H., Schmidt, J.B., Silva, I., Szuts, T., Castro, A. & Crespo, L.C. (2008) Rapid biodiversity assessment of spiders (Araneae) using semi-quantitative sampling: a case study in a Mediterranean forest. Insect Conservation and Diversity, 1: 71-84.
}
\keyword{datasets}

