% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BAT.R
\docType{data}
\name{phylotree}
\alias{phylotree}
\title{Taxonomic tree for 338 species of spiders (surrogate for phylogeny)}
\format{
An hclust object with 338 species.
}
\usage{
data(phylotree)
}
\description{
A dataset representing an approximation to the phylogenetic tree for 338 species of spiders captured in Portugal.
The tree is based on the linnean hierarchy, with different suborders separated by 1 unit, families by 0.75, genera by 0.5 and species by 0.25.
}
\keyword{datasets}
