% Generated by roxygen2 (4.0.2): do not edit by hand
\name{sim.plot}
\alias{sim.plot}
\title{Plots of simulated species spatial distributions.}
\usage{
sim.plot(comm, sad = FALSE, n = 0)
}
\arguments{
\item{comm}{artificial community data from function sim.spatial.}

\item{sad}{boolean indicating if the SAD plot should also be shown. Default is FALSE.}

\item{n}{number of species to plot simultaneously. Default is the number of species in comm.}
}
\description{
Plots individuals from artificial communities with given SAD and spatial clustering.
}
\details{
Function useful for visualizing the results of sim.spatial.
}
\examples{
comm <- sim.spatial(1000, 24)
sim.plot(comm)
sim.plot(comm, sad = TRUE)
sim.plot(comm, n = 9)
}

