% Generated by roxygen2 (4.0.2): do not edit by hand
\name{sim.tree}
\alias{sim.tree}
\title{Simulated phylogenetic or functional tree.}
\usage{
sim.tree(n, m = 2)
}
\arguments{
\item{n}{number of species.}

\item{m}{number of genes/traits. Default is 2.}
}
\value{
An hclust object.
}
\description{
Simulates a random tree.
}
\details{
A very simple tree based on random genes/traits.
}
\examples{
tree <- sim.tree(100)
plot(as.dendrogram(tree))
}

