% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BAT.R
\name{dummy}
\alias{dummy}
\title{Dummify variables.}
\usage{
dummy(trait)
}
\arguments{
\item{trait}{a species x traits matrix or data.frame.}
}
\value{
A matrix with variables converted when needed.
}
\description{
Convert factor variables to dummy variables.
}
\examples{
trait = data.frame(length = c(2,4,6,3,1), wing = c("A", "B", "A", "A", "B"))
dummy(trait)
}
