% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BAT.R
\docType{data}
\name{geres}
\alias{geres}
\title{Sample data of spiders in Geres (Portugal)}
\format{
A data frame with 320 sampling untis (rows) and 338 species (variables).
}
\usage{
data(geres)
}
\description{
A dataset containing the abundance of 338 spider species in each of 320 sampling units. Details are described in:
Cardoso, P., Scharff, N., Gaspar, C., Henriques, S.S., Carvalho, R., Castro, P.H., Schmidt, J.B., Silva, I., Szuts, T., Castro, A. & Crespo, L.C. (2008) Rapid biodiversity assessment of spiders (Araneae) using semi-quantitative sampling: a case study in a Mediterranean forest. Insect Conservation and Diversity, 1: 71-84.
}
\keyword{datasets}
