% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BAT.R
\docType{data}
\name{guadiana}
\alias{guadiana}
\title{Sample data of spiders in Guadiana (Portugal)}
\format{
A data frame with 192 sampling units (rows) and 338 species (variables).
}
\usage{
data(guadiana)
}
\description{
A dataset containing the abundance of 338 spider species in each of 320 sampling units. Details are described in:
Cardoso, P., Henriques, S.S., Gaspar, C., Crespo, L.C., Carvalho, R., Schmidt, J.B., Sousa, P. & Szuts, T. (2009) Species richness and composition assessment of spiders in a Mediterranean scrubland. Journal of Insect Conservation, 13: 45-55.
}
\keyword{datasets}
